One client class only.
